NET-105: fix cycle imports
