+++ api/utils.go
@@ -1,0 +1,45 @@
+/*
+ *    Copyright 2018 Insolar
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package api
+
+import (
+	"crypto"
+
+	"github.com/insolar/insolar/component"
+	"github.com/insolar/insolar/core"
+	"github.com/insolar/insolar/cryptography"
+	"github.com/insolar/insolar/platformpolicy"
+)
+
+type inPlaceKeyStore struct {
+	privateKey crypto.PrivateKey
+}
+
+func (ipks *inPlaceKeyStore) GetPrivateKey(string) (crypto.PrivateKey, error) {
+	return ipks.privateKey, nil
+}
+
+func NewCryptographyService(privateKey crypto.PrivateKey) core.CryptographyService {
+	cs := platformpolicy.NewPlatformCryptographyScheme()
+	ks := &inPlaceKeyStore{privateKey: privateKey}
+	kp := platformpolicy.NewKeyProcessor()
+	service := cryptography.NewCryptographyService()
+	cm := component.Manager{}
+	cm.Register(cs)
+	cm.Inject(ks, service, kp)
+	return service
+}
